+++ CodeStarterCamp_Week4/main.swift
@@ -1,6 +1,12 @@
-
 import Foundation
 
+
+func printResult(activity: Activity) {
+    activity.action(bodyCondition)
+    bodyCondition.printCondtion()
+}
+
+var bodyCondition = BodyCondition()
 printResult(activity: squart)
